requirements.txt
Flask==1.1.2
python-dotenv==0.15.0
google-cloud-datastore==1.15.3
google-auth==1.22.1
requests==2.24.0